Added requirements.txt
Updated code
